% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WinBox.R
\name{WinBox}
\alias{WinBox}
\alias{closeWinBox}
\title{WinBox}
\usage{
WinBox(
  title,
  ui,
  options = optionsWinBox(),
  id = NULL,
  session = shiny::getDefaultReactiveDomain()
)

closeWinBox(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title}{Title for the window.}

\item{ui}{Content of the window.}

\item{options}{List of options, see \code{\link[=optionsWinBox]{optionsWinBox()}}.}

\item{id}{An unique identifier for the window.}

\item{session}{Shiny session.}
}
\value{
No value.
}
\description{
A window manager with JavaScript library \href{https://nextapps-de.github.io/winbox/}{WinBox.js}.
}
